requirements.txt
fastapi
uvicorn[standard]
google-analytics-data
pydantic
python-dotenv
